+++ src/main/java/com/twilio/rest/api/v2010/account/incomingphonenumber/Local.java
@@ -597,9 +597,9 @@ public class Local extends Resource {
     }
 
     /**
-     * Returns Whether the phone number is enabled for emergency calling.
+     * Returns Displays if emergency calling is enabled for this number..
      *
-     * @return Whether the phone number is enabled for emergency calling
+     * @return Displays if emergency calling is enabled for this number.
      */
     public final Local.EmergencyStatus getEmergencyStatus() {
         return this.emergencyStatus;