+++ src/main/java/com/twilio/rest/api/v2010/account/incomingphonenumber/TollFree.java
@@ -600,9 +600,9 @@ public class TollFree extends Resource {
     }
 
     /**
-     * Returns Whether the phone number is enabled for emergency calling.
+     * Returns Displays if emergency calling is enabled for this number..
      *
-     * @return Whether the phone number is enabled for emergency calling
+     * @return Displays if emergency calling is enabled for this number.
      */
     public final TollFree.EmergencyStatus getEmergencyStatus() {
         return this.emergencyStatus;