+++ src/main/java/com/twilio/rest/verify/v2/service/VerificationCreator.java
@@ -89,12 +89,13 @@ public class VerificationCreator extends Creator<Verification> {
     }
 
     /**
-     * The locale to use for the verification SMS or call. Can be: `af`, `ar`, `ca`,
-     * `cs`, `da`, `de`, `el`, `en`, `en-GB`, `es`, `fi`, `fr`, `he`, `hi`, `hr`,
-     * `hu`, `id`, `it`, `ja`, `ko`, `ms`, `nb`, `nl`, `pl`, `pt`, `pr-BR`, `ro`,
-     * `ru`, `sv`, `th`, `tl`, `tr`, `vi`, `zh`, `zh-CN`, or `zh-HK.`.
+     * The locale to use for the verification SMS, WhatsApp or call. Can be: `af`,
+     * `ar`, `ca`, `cs`, `da`, `de`, `el`, `en`, `en-GB`, `es`, `fi`, `fr`, `he`,
+     * `hi`, `hr`, `hu`, `id`, `it`, `ja`, `ko`, `ms`, `nb`, `nl`, `pl`, `pt`,
+     * `pr-BR`, `ro`, `ru`, `sv`, `th`, `tl`, `tr`, `vi`, `zh`, `zh-CN`, or
+     * `zh-HK.`.
      *
-     * @param locale The locale to use for the verification SMS or call
+     * @param locale The locale to use for the verification SMS, WhatsApp or call
      * @return this
      */
     public VerificationCreator setLocale(final String locale) {